# Relyf.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Relyf API",
    "description": "API for the Relyf upcycling platform - connecting users with creative ways to reuse and repurpose items",
    "version": "v1",
    "contact": {
      "name": "Relyf Team",
      "url": "https://github.com/nanasarf/Relyf"
    }
  },
  "servers": [
    {
      "url": "https://localhost:7139",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/Users/check-username/{userName}": {
      "get": {
        "tags": ["Users"],
        "summary": "Check if username is available",
        "description": "Validates username format and checks availability",
        "operationId": "CheckUsername",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          }
        ],
        "responses": {
          "200": {
            "description": "Username check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUsernameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user profile by ID",
        "description": "Retrieve detailed user profile including follower counts and relationship status",
        "operationId": "GetUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Update user profile information (requires authentication and ownership)",
        "operationId": "UpdateUserProfile",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 123
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              },
              "example": {
                "userName": "john_doe_updated",
                "displayName": "John Doe Updated",
                "bio": "I love upcycling and sustainability!",
                "avatarUrl": "https://example.com/avatar.jpg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (validation error)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Username can only contain letters, numbers, and underscores."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Cannot update other user's profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Cannot update another user's profile"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "User not found"
                }
              }
            }
          },
          "409": {
            "description": "Username already taken",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Username is already taken."
                }
              }
            }
          }
        }
      }
    },
    "/api/Users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users",
        "description": "Search users by username, display name, or email",
        "operationId": "SearchUsers",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "john"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSearchResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Users/{id}/followers": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user followers",
        "description": "List all users following this user",
        "operationId": "GetFollowers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of followers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Users/{id}/following": {
      "get": {
        "tags": ["Users"],
        "summary": "Get users that this user follows",
        "description": "List all users this user is following",
        "operationId": "GetFollowing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users being followed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfile"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckUsernameResponse": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "pattern": "^[a-zA-Z0-9_]+$",
            "nullable": true,
            "description": "New username (optional, must be unique)",
            "example": "john_doe_updated"
          },
          "displayName": {
            "type": "string",
            "maxLength": 120,
            "nullable": true,
            "description": "New display name (optional)",
            "example": "John Doe Updated"
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "User biography (optional)",
            "example": "I love upcycling and sustainability!"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "Avatar image URL (optional)",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "description": "All fields are optional - only provided fields will be updated"
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "followerCount": {
            "type": "integer"
          },
          "followingCount": {
            "type": "integer"
          },
          "projectCount": {
            "type": "integer"
          },
          "ideaCount": {
            "type": "integer"
          },
          "isFollowing": {
            "type": "boolean",
            "description": "Whether the requesting user is following this user"
          },
          "isFollowedBy": {
            "type": "boolean",
            "description": "Whether this user is following the requesting user"
          }
        }
      },
      "UserSearchResult": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserProfile"
            }
          },
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "take": {
            "type": "integer"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    }
  },
...
}